Updated Editing Post for #10

+++ src/models/collection.ts
@@ -46,16 +46,16 @@ export class Collection extends Model {
   @Column("updated", TEXT)
   public updated: Date = null;
 
-  @Column("can_read", DOUBLE)
+  @Column("can_read", INTEGER)
   public can_read: boolean = null;
 
-  @Column("can_create", DOUBLE)
+  @Column("can_create", INTEGER)
   public can_create: boolean = null;
 
-  @Column("can_update", DOUBLE)
+  @Column("can_update", INTEGER)
   public can_update: boolean = null;
 
-  @Column("can_delete", DOUBLE)
+  @Column("can_delete", INTEGER)
   public can_delete: boolean = null;
 
 }
